+++ packages/fabrik_theme/lib/fabrik_theme.dart
@@ -1,4 +1,4 @@
-export 'src/responsive/responsive.dart';
+export 'src/builders/builders.dart';
+export 'src/extensions/extensions.dart';
 export 'src/theme/theme.dart';
 export 'src/tokens/tokens.dart';
-export 'src/typography/typography.dart';