+++ packages/fabrik_theme/lib/src/responsive/responsive.dart
@@ -1,2 +1,0 @@
-export 'fabrik_layout.dart';
-export 'fabrik_responsive.dart';